requirements.txt
Flask==3.0.3
joblib==1.4.2
matplotlib==3.9.2
numpy
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.5.1
scipy==1.14.1
tensorflow
ultralytics==8.2.64
Werkzeug
